Change design: Replace `Slice<Marker>` with `Marker`
This is necessary for the code generation of structure types to work.  The
code will be generated by the `define_gv!()` macro, and the types will be
spat out in the crate that is using them, rather than in this crate.
We can't just have the macro spit out a marker because we need to define
inherent impls and it's not valid to define an inherent impl for a struct
in a different crate.