macro: Generalise item offsets
This generated impl will panic on non-normal inputs, but it's a step
forward.  I've copied the algorithm from the `GVariant` paper.  See there
for documentation on what it all means.

The return type of `.split` has changed here.  We return tuple of
`&impl GVariantMarker`, one for each field.  Previously we'd return rust
types.  So for `(si)` it's now:

    (&marker::S, &marker::I)

rather than the more convenient

    (&[u8], &i32)

The new return type is more general and requires checking fewer of the
invariants required for constructing the type than before, but it is less
convenient.  I'll have to think about if we can do better here.