Redesign: Replace marker types with types with same repr as GVariant
This is nicer because the user doesn't have to convert from the marker type
to the real type every time.  So before it was:

    let i = <gv!("i")>::mark(data).to_rs();

it's now:

    let i = <gv!("i")>::from_aligned_bytes(data);

The advantages are more significant with generated structs.  For example:
`.to_tuple()` for **(iy)** used to return `(&marker::<I>, &marker<Y>)`,
while now it returns `(&i32, &u8)`.

I've had to split maybe types and array types into sized and non-sized.
This is fine as the user shouldn't be naming these types anyway, but should
be using `gv!` instead.  This has worked out well because an array of a
`Sized` type is now just a rust slice (`[T]`).