macro: Use macro_rules! and proc macro in tandem to get hygiene
This is nice because the user no longer needs to use `define_gv!`.  All the
types are defined into a private module thanks to `macro_rules!`.  We also
use the `Marker` trait to enable "passing through" the typestr as a
`&[u8]`.  This will be useful when it comes to implementing variants.  We
will have a single type we can use to refer to the (type, typestr) pair -
and thus be able to extract a value of the right type from a Variant, but
only if the typestr matches.