Improve documentation and conformance of `Str` type
I've changed `to_bytes()` to be spec-conformant and added a
`to_bytes_non_conformant` for users who aren't happy with the trade-off.
There could potentially be security implications of having the same data
interpreted differently by different implementations so the default should
be to conform and let people opt out.

In practice if the user is planning on doing something with the data and
it fits in cache then the performance impact of checking for NULs should be
~0.