Use the memchr crate for finding NUL bytes
This is easier to use and faster than our implementation using
`iter().position()`.

I believe the [memchr] crate to be reliable as it's authored by
@BurntSushi, and has many significant reverse dependencies including regex,
tokio and nom.

It only needs to depend on std for CPU feature detection to enable AVX
optimisations, so we propogate our std feature to it.

[memchr]: https://github.com/BurntSushi/rust-memchr