Fix deserialisation of some structure types
If `c < B` and `x` is aligned to `B` then we know that none of the bits
in `x` are set in `c`.  This makes `x | c == x + c`.  The GVariant spec
includes a transformation that makes `c < B`, but we haven't applied this
transformation, so we don't have this guarantee.  This caused some GVariant
structures to be deserialised incorrectly.

It applies to certain complex structures with multiple fixed-size types
following non-fixed size types.  Exactly which I'm not sure, but this is
obviously a serious issue.

This commit fixes that issue.  This was discovered by my new struct fuzzing
test.  I've added an explicit regression test.