Replace trait Marker with struct Marker
Moving the relevant typestr handling into `trait Cast`.  This is a little
simpler as we don't have to generate 2 types every time we call `gv!` and
users no longer need to import trait Marker, all the methods are on
`struct Marker` directly.  Types now know what the equivalent GVariant type
is as part of `Cast`.