Add new struct GString
Replaces `Box<Str>` as the owned version of `Str`.  It's easier to work
with and more flexible.  We can easily and cheaply convert to/from
`std::string::String`.  We can append to it and use the `write!` macro.

Unlike with `Str`, the internal representation is always in normal form.

This is a breaking change.