Teach `AlignedTo` that 8B is GVariant's maximum alignment
This says that 8B is aligned to all other alignments.  Previously the other
alignments were listed explicitly, but rust doesn't know that there isn't
some A16 lurking in another crate somewhere, so that wasn't enough.

This simplifies trait bounds.  A generic function can now receive an
`AlignedSlice<A8>` and generically cast it to any other `AlignedSlice<...>`
with `.as_aligned()`.  This will be useful in the next few commits that
introduce an owned version of `AlignedSlice<A8>`.