Add new class `AlignedBuf`
This is like a `Vec<u8>`, but explicitly maintains the invariant that the
data is 8B aligned.  In practice `Vec<u8>` are always 8B aligned anyway[^1]
(except when using miri), which makes conversions to/from `Vec<u8>` cheap.

This will make it much easier to integrate with the rest of the rust
ecosystem.

[^1]: Except when running under miri